requirements.txt
streamlit
openai==1.12.0
gunicorn
python-dotenv
